// src/pages/abouts-us/enum.ts
enum TagNames {
  DIV = 'div',
  NAV = 'nav',
  A = 'a',
  IMG = 'img',
  SPAN = 'span',
  H1 = 'h1',
  H2 = 'h2',
  H3 = 'h3',
  P = 'p',
}

enum Styles {
  CONTAINER = 'about-us-container',
  PROJ_DESCRIPTION = 'project-description',
  DEVCARD = 'devcard',
  HEADER = 'header',
  SUBHEADER = 'subheader',
  DESCRIPTION_LEFT = 'description-left',
  DESCRIPTION_RIGHT = 'description-right',
  DESCRIPTION_BOTTOM = 'description-bottom',
  PARAGRAPH = 'paragraph',
  DEV_CARD = 'developer-card',
  PHOTO_CONTAINER = 'photo-container',
  PHOTO = 'photo',
  DESCRIPTION_CONTAINER = 'description-container',
  CARD_HEADING = 'card-heading',
  CARD_POSITION = 'card-position',
  CARD_DESCRIPTION = 'card-description',
  GH_LINK = 'github-link',
  RS_LINK = 'rs-link',
  RS_IMG = 'rs-img',
}

enum Attributes {
  SRC = 'src',
  HREF = 'href',
  ALT = 'alt',
  TARGET = 'target',
}

export { TagNames, Styles, Attributes };